syscall: handle special mips64 dirent and regenerate mips ztypes
Modify types_linux.go so that the ztypes_linux_mips64* files can be
regenerated, then regenerate all the ztypes_linux_mips* files. The
mips64 files files previously had some manual modifications made to
struct dirent so that the getdents syscall can be used instead of the
getdents64 syscall.

As a side effect, the mips64* Stat_t now uses the n64 version of
struct stat instead of the o32 version it has previously been using.

+++ src/syscall/syscall_linux_mips64x.go
@@ -168,18 +168,18 @@ func Stat(path string, s *Stat_t) (err error) {
 }
 
 func fillStat_t(s *Stat_t, st *stat_t) {
-	s.Dev = st.Dev
+	s.Dev = uint64(st.Dev)
 	s.Ino = st.Ino
 	s.Mode = st.Mode
-	s.Nlink = st.Nlink
+	s.Nlink = uint64(st.Nlink)
 	s.Uid = st.Uid
 	s.Gid = st.Gid
-	s.Rdev = st.Rdev
+	s.Rdev = uint64(st.Rdev)
 	s.Size = st.Size
 	s.Atim = Timespec{int64(st.Atime), int64(st.Atime_nsec)}
 	s.Mtim = Timespec{int64(st.Mtime), int64(st.Mtime_nsec)}
 	s.Ctim = Timespec{int64(st.Ctime), int64(st.Ctime_nsec)}
-	s.Blksize = st.Blksize
+	s.Blksize = int64(st.Blksize)
 	s.Blocks = st.Blocks
 }
